requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9
pandas==2.2.1
scikit-learn==1.4.1.post1
numpy==1.26.4